+++ XenAdmin/Commands/DragDropCrossPoolMoveHaltedVMCommand.cs
@@ -73,7 +73,7 @@ namespace XenAdmin.Commands
 
                     var homeHost = vm.Home();
 
-                    if (Helpers.productVersionCompare(Helpers.HostProductVersion(targetHost),
+                    if (Helpers.ProductVersionCompare(Helpers.HostProductVersion(targetHost),
                             Helpers.HostProductVersion(homeHost ?? Helpers.GetCoordinator(vm.Connection))) < 0)
                         return Messages.OLDER_THAN_CURRENT_SERVER;
 
@@ -120,7 +120,7 @@ namespace XenAdmin.Commands
                 if (vm.allowed_operations == null || !vm.allowed_operations.Contains(vm_operations.migrate_send))
                     return false;
 
-                if (Helpers.productVersionCompare(Helpers.HostProductVersion(targetHost), Helpers.HostProductVersion(homeHost ?? Helpers.GetCoordinator(vm.Connection))) < 0)
+                if (Helpers.ProductVersionCompare(Helpers.HostProductVersion(targetHost), Helpers.HostProductVersion(homeHost ?? Helpers.GetCoordinator(vm.Connection))) < 0)
                     return false;
 
                 if (VMOperationHostCommand.VmCpuIncompatibleWithHost(targetHost, vm))
