+++ XenModel/Utils/Helpers.cs
@@ -45,7 +45,7 @@ using XenCenterLib;
 
 namespace XenAdmin.Core
 {
-    public static class Helpers
+    public static partial class Helpers
     {
         private static readonly log4net.ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
@@ -59,9 +59,6 @@ namespace XenAdmin.Core
         public const string GuiTempObjectPrefix = "__gui__";
 
         public static NumberFormatInfo _nfi = new CultureInfo("en-US", false).NumberFormat;
-
-        public static readonly Regex SessionRefRegex = new Regex(@"OpaqueRef:[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}");
-
 
         /// <summary>
         /// Return the given host's product version, or the pool coordinator's product version if
@@ -105,82 +102,6 @@ namespace XenAdmin.Core
             }
 
             return output;
-        }
-
-        /// <summary>
-        /// Only log the unrecognised version message once (CA-11201).
-        /// </summary>
-        private static bool _unrecognisedVersionWarned = false;
-        /// <summary>
-        /// Numbers should have three parts, i.e. be in the form a.b.c, otherwise they won't be parsed.
-        /// </summary>
-        /// <param name="version1">May be null.</param>
-        /// <param name="version2">May be null.</param>
-        /// <returns></returns>
-        public static int productVersionCompare(string version1, string version2)
-        {
-            // Assume version numbers are of form 'a.b.c'
-            int a1 = 99, b1 = 99, c1 = 99, a2 = 99, b2 = 99, c2 = 99;
-
-
-            string[] tokens = null;
-            if (version1 != null)
-                tokens = version1.Split('.');
-            if (tokens != null && tokens.Length == 3)
-            {
-                a1 = int.Parse(tokens[0]);
-                b1 = int.Parse(tokens[1]);
-                c1 = int.Parse(tokens[2]);
-            }
-            else
-            {
-                if (!_unrecognisedVersionWarned)
-                {
-                    log.DebugFormat("Unrecognised version format {0} - treating as developer version", version1);
-                    _unrecognisedVersionWarned = true;
-                }
-            }
-            tokens = null;
-            if (version2 != null)
-                tokens = version2.Split('.');
-            if (tokens != null && tokens.Length == 3)
-            {
-                a2 = int.Parse(tokens[0]);
-                b2 = int.Parse(tokens[1]);
-                c2 = int.Parse(tokens[2]);
-            }
-            else
-            {
-                if (!_unrecognisedVersionWarned)
-                {
-                    log.DebugFormat("Unrecognised version format {0} - treating as developer version", version2);
-                    _unrecognisedVersionWarned = true;
-                }
-            }
-
-            if (a2 > a1)
-            {
-                return -1;
-            }
-            else if (a2 == a1)
-            {
-                if (b2 > b1)
-                {
-                    return -1;
-                }
-                else if (b2 == b1)
-                {
-                    if (c2 > c1)
-                    {
-                        return -1;
-                    }
-                    else if (c1 == c2)
-                    {
-                        return 0;
-                    }
-                }
-            }
-            return 1;
         }
 
         /// <summary>
@@ -274,276 +195,6 @@ namespace XenAdmin.Core
             }
             return false;
         }
-
-        #region Versions
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool DundeeOrGreater(IXenConnection conn)
-        {
-            return conn == null || DundeeOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Dundee is ver. 2.0.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool DundeeOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.0.0") >= 0;
-        }
-
-        public static bool DundeePlusOrGreater(IXenConnection conn)
-        {
-            return conn == null || conn.Session == null || conn.Session.APIVersion >= API_Version.API_2_6;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool ElyOrGreater(IXenConnection conn)
-        {
-            return conn == null || ElyOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Ely is ver. 2.1.1
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool ElyOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.1.1") >= 0;
-        }
-
-        public static bool HavanaOrGreater(IXenConnection conn)
-        {
-            return conn == null || HavanaOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// As Havana platform version is same with Ely and Honolulu, so use product version here
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool HavanaOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-            string product_version = HostProductVersion(host);
-            return
-                product_version != null &&
-                ElyOrGreater(host) &&
-                !FalconOrGreater(host) &&
-                productVersionCompare(product_version, BrandManager.ProductVersion712) >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool FalconOrGreater(IXenConnection conn)
-        {
-            return conn == null || FalconOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Falcon is ver. 2.3.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool FalconOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.2.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool InvernessOrGreater(IXenConnection conn)
-        {
-            return conn == null || InvernessOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Inverness is ver. 2.4.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool InvernessOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.3.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool JuraOrGreater(IXenConnection conn)
-        {
-            return conn == null || JuraOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Jura is ver. 2.5.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool JuraOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.4.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool KolkataOrGreater(IXenConnection conn)
-        {
-            return conn == null || KolkataOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Kolkata platform version is 2.6.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool KolkataOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.5.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool LimaOrGreater(IXenConnection conn)
-        {
-            return conn == null || LimaOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// Lima platform version is 2.7.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool LimaOrGreater(Host host)
-        {
-            if (host == null)
-                return true;
-
-            string platform_version = HostPlatformVersion(host);
-            return platform_version != null && productVersionCompare(platform_version, "2.6.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool NaplesOrGreater(IXenConnection conn)
-        {
-            return conn == null || NaplesOrGreater(GetCoordinator(conn));
-        }
-
-        /// Naples is ver. 3.0.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool NaplesOrGreater(Host host)
-        {
-            return host == null || NaplesOrGreater(HostPlatformVersion(host));
-        }
-
-        public static bool NaplesOrGreater(string platformVersion)
-        {
-            return platformVersion != null && productVersionCompare(platformVersion, "2.9.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool QuebecOrGreater(IXenConnection conn)
-        {
-            return conn == null || QuebecOrGreater(GetCoordinator(conn));
-        }
-
-        /// Quebec platform version is 3.1.0
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool QuebecOrGreater(Host host)
-        {
-            return host == null || QuebecOrGreater(HostPlatformVersion(host));
-        }
-
-        public static bool QuebecOrGreater(string platformVersion)
-        {
-            return platformVersion != null && productVersionCompare(platformVersion, "3.0.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool StockholmOrGreater(IXenConnection conn)
-        {
-            return conn == null || StockholmOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool StockholmOrGreater(Host host)
-        {
-            return host == null || StockholmOrGreater(HostPlatformVersion(host));
-        }
-
-        /// Stockholm is ver. 3.2.0
-        public static bool StockholmOrGreater(string platformVersion)
-        {
-            return platformVersion != null && productVersionCompare(platformVersion, "3.1.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool Post82X(IXenConnection conn)
-        {
-            return conn == null || Post82X(GetCoordinator(conn));
-        }
-
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool Post82X(Host host)
-        {
-            return host == null || Post82X(HostPlatformVersion(host));
-        }
-
-        public static bool Post82X(string platformVersion)
-        {
-            return platformVersion != null && productVersionCompare(platformVersion, "3.2.50") >= 0;
-        }
-
-        /// <param name="conn">May be null, in which case true is returned.</param>
-        public static bool YangtzeOrGreater(IXenConnection conn)
-        {
-            return conn == null || YangtzeOrGreater(Helpers.GetCoordinator(conn));
-        }
-
-        /// <param name="host">May be null, in which case true is returned.</param>
-        public static bool YangtzeOrGreater(Host host)
-        {
-            return host == null || YangtzeOrGreater(HostPlatformVersion(host));
-        }
-
-        /// Yangtze is ver. 3.2.1
-        public static bool YangtzeOrGreater(string platformVersion)
-        {
-            return platformVersion != null && productVersionCompare(platformVersion, "3.2.1") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_1_290_0(IXenConnection conn)
-        {
-            var coordinator = GetCoordinator(conn);
-            return coordinator == null || productVersionCompare(coordinator.GetXapiVersion(), "1.290.0") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_1_290_0(Host host)
-        {
-            return host == null || productVersionCompare(host.GetXapiVersion(), "1.290.0") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_1_313_0(IXenConnection conn)
-        {
-            var coordinator = GetCoordinator(conn);
-            return coordinator == null || productVersionCompare(coordinator.GetXapiVersion(), "1.313.0") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_1_313_0(Host host)
-        {
-            return host == null || productVersionCompare(host.GetXapiVersion(), "1.313.0") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_22_5_0(IXenConnection conn)
-        {
-            var coordinator = GetCoordinator(conn);
-            return coordinator == null || productVersionCompare(coordinator.GetXapiVersion(), "22.5.0") >= 0;
-        }
-
-        public static bool XapiEqualOrGreater_22_19_0(IXenConnection conn)
-        {
-            var coordinator = GetCoordinator(conn);
-            return coordinator == null || productVersionCompare(coordinator.GetXapiVersion(), "22.19.0") >= 0;
-        }
-
-        #endregion
 
         // CP-3435: Disable Check for Updates in Common Criteria Certification project
         public static bool CommonCriteriaCertificationRelease
@@ -784,7 +435,7 @@ namespace XenAdmin.Core
             Host coordinator = Helpers.GetCoordinator(host.Connection);
             if (coordinator == null || coordinator.opaque_ref == host.opaque_ref)
                 return false;
-            else if (Helpers.productVersionCompare(Helpers.HostProductVersion(host), Helpers.HostProductVersion(coordinator)) >= 0)
+            else if (Helpers.ProductVersionCompare(Helpers.HostProductVersion(host), Helpers.HostProductVersion(coordinator)) >= 0)
                 return false;
             else
                 return true;
