+++ XenModel/XenModel.csproj
@@ -180,6 +180,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>UnitStrings.resx</DependentUpon>
     </Compile>
+    <Compile Include="Utils\Helpers.Versions.cs" />
     <Compile Include="Utils\StringExtensions.cs" />
     <Compile Include="Utils\TimeSpanExtensions.cs" />
     <Compile Include="WLB\WlbPoolConfiguration.cs" />
